guard dynamic string repeat allocation

+++ crates/robin-sparkless-polars/src/udfs/rest.rs
@@ -177,7 +177,10 @@ pub fn apply_repeat_dynamic(columns: &mut [Column]) -> PolarsResult<Option<Colum
         name.as_str().into(),
         (0..out_len).map(|idx| match (value_at(idx), count_at(idx)) {
             (Some(s), Some(n)) if n > 0 => usize::try_from(n).ok().and_then(|count| {
-                if !s.is_empty() && count > usize::MAX / s.len() {
+                // `String::repeat` rejects lengths above `isize::MAX`, which can be
+                // smaller than `usize::MAX` on 64-bit targets. Check that limit here
+                // so an untrusted count produces Spark's NULL rather than panicking.
+                if !s.is_empty() && count > isize::MAX as usize / s.len() {
                     None
                 } else {
                     Some(s.repeat(count))
@@ -858,8 +861,8 @@ pub fn apply_arrays_zip(columns: &mut [Column]) -> PolarsResult<Option<Column>> 
         .collect::<PolarsResult<Vec<_>>>()?;
     let len = list_cas[0].len();
     let inner_dtype = list_cas[0].inner_dtype().clone();
-    use polars::chunked_array::builder::get_list_builder;
     use polars::chunked_array::StructChunked;
+    use polars::chunked_array::builder::get_list_builder;
     use polars::datatypes::Field;
     let struct_fields: Vec<Field> = (0..n)
         .map(|i| Field::new(format!("field_{i}").into(), inner_dtype.clone()))
@@ -1042,8 +1045,8 @@ pub fn apply_str_to_map(
     pair_delim: &str,
     key_value_delim: &str,
 ) -> PolarsResult<Option<Column>> {
-    use polars::chunked_array::builder::get_list_builder;
     use polars::chunked_array::StructChunked;
+    use polars::chunked_array::builder::get_list_builder;
     use polars::datatypes::Field;
     let name = column.field().into_owned().name;
     let series = column.take_materialized_series();
@@ -1097,8 +1100,8 @@ pub fn apply_map_concat(
     columns: &mut [Column],
     disable_map_key_normalization: bool,
 ) -> PolarsResult<Option<Column>> {
-    use polars::chunked_array::builder::get_list_builder;
     use polars::chunked_array::StructChunked;
+    use polars::chunked_array::builder::get_list_builder;
     use polars::datatypes::Field;
     if columns.len() < 2 {
         return Err(PolarsError::ComputeError(
@@ -2086,9 +2089,9 @@ pub fn apply_try_to_binary(column: Column, fmt: &str) -> PolarsResult<Option<Col
 
 /// AES-GCM encrypt (PySpark aes_encrypt). Key: UTF-8 string, 16 or 32 bytes for AES-128/256. Output: hex(nonce||ciphertext).
 fn aes_gcm_encrypt_one(plaintext: &[u8], key: &[u8]) -> Option<String> {
+    use aes_gcm::Aes128Gcm;
     use aes_gcm::aead::generic_array::GenericArray;
     use aes_gcm::aead::{Aead, KeyInit};
-    use aes_gcm::Aes128Gcm;
     use rand::RngCore;
     let key_arr: [u8; 16] = key
         .iter()
@@ -2110,9 +2113,9 @@ fn aes_gcm_encrypt_one(plaintext: &[u8], key: &[u8]) -> Option<String> {
 
 /// AES-GCM decrypt (PySpark aes_decrypt). Input: hex(nonce||ciphertext).
 fn aes_gcm_decrypt_one(hex_input: &str, key: &[u8]) -> Option<String> {
+    use aes_gcm::Aes128Gcm;
     use aes_gcm::aead::generic_array::GenericArray;
     use aes_gcm::aead::{Aead, KeyInit};
-    use aes_gcm::Aes128Gcm;
     let bytes = hex::decode(hex_input.as_bytes()).ok()?;
     if bytes.len() < 12 + 16 {
         return None; // nonce + at least tag
@@ -3346,8 +3349,8 @@ pub fn apply_map_from_arrays(
     columns: &mut [Column],
     disable_map_key_normalization: bool,
 ) -> PolarsResult<Option<Column>> {
-    use polars::chunked_array::builder::get_list_builder;
     use polars::chunked_array::StructChunked;
+    use polars::chunked_array::builder::get_list_builder;
     use polars::datatypes::Field;
     if columns.len() < 2 {
         return Err(PolarsError::ComputeError(
@@ -3420,8 +3423,8 @@ pub fn apply_map_from_arrays(
 
 /// Zip two array columns into List(Struct{left, right}) for zip_with. Shorter padded with null.
 pub fn apply_zip_arrays_to_struct(columns: &mut [Column]) -> PolarsResult<Option<Column>> {
-    use polars::chunked_array::builder::get_list_builder;
     use polars::chunked_array::StructChunked;
+    use polars::chunked_array::builder::get_list_builder;
     use polars::datatypes::Field;
     if columns.len() < 2 {
         return Err(PolarsError::ComputeError(
@@ -3539,8 +3542,8 @@ pub fn apply_zip_arrays_to_struct(columns: &mut [Column]) -> PolarsResult<Option
 
 /// Merge two maps into List(Struct{key, value1, value2}) for map_zip_with. Union of keys.
 pub fn apply_map_zip_to_struct(columns: &mut [Column]) -> PolarsResult<Option<Column>> {
-    use polars::chunked_array::builder::get_list_builder;
     use polars::chunked_array::StructChunked;
+    use polars::chunked_array::builder::get_list_builder;
     use polars::datatypes::Field;
     use std::collections::BTreeMap;
     if columns.len() < 2 {
@@ -5063,7 +5066,7 @@ pub fn apply_url_decode(column: Column) -> PolarsResult<Option<Column>> {
 
 /// url_encode(column) - percent-encode string for URL (PySpark url_encode).
 pub fn apply_url_encode(column: Column) -> PolarsResult<Option<Column>> {
-    use percent_encoding::{utf8_percent_encode, NON_ALPHANUMERIC};
+    use percent_encoding::{NON_ALPHANUMERIC, utf8_percent_encode};
     let name = column.field().into_owned().name;
     let series = column.take_materialized_series();
     let ca = series.str().map_err(|e| compute_err("url_encode", e))?;
